Move task list components (TaskItem, TaskList) into self-contained folde
