+++ src/app.tsx
@@ -5,7 +5,7 @@ import { useTheme } from './hooks/use-theme'
 import { useHashRoute } from './hooks/use-hash-route'
 import { TaskInput } from './components/task-input/task-input'
 import { Filters } from './components/filters/filters'
-import { TaskList } from './components/task-list'
+import { TaskList } from './components/task-list/task-list'
 import { Footer } from './components/footer/footer'
 import { StatsPage } from './components/stats-page'
 import { NavLink } from './components/nav-link/nav-link'
