+++ src/components/task-item/task-item.tsx
@@ -1,4 +1,4 @@
-import { Task } from '../types'
+import { Task } from '../../types'
 
 interface TaskItemProps {
   task: Task
